Modify original utilities.notice() method
Now it can print notices in a gui instead of the console if specified
when instantiating.
However, every call of such method needs to be modified because it now
takes one more argument. Besides, a 'gui' attribute must be passed to
every object that calls notice() and that gui must have a
gui.write_debugger('str') method.
If no gui is specified, it prints to console, as always.