Reorganize project's structure
A setup-py file was added, because of that source files were moved to
py3gestalt. By the way, the name changed officially to 'py3gestalt'
(lowercase).
All html content was moved to a sub directory and an index.html in
'docs' redirects to the original index.html.
GUI was moved to examples.